% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AISsub.R
\docType{data}
\name{AISsub}
\alias{AISsub}
\title{AISsub}
\format{
A data frame with 202 rows (each athlete) and 13 columns:
\describe{
  \item{Sex}{The sex of the athlete; one of \code{F} or \code{M}}
  \item{Sport}{The sport played by the athlete; one of \code{BBall} (basketball), \code{Field}, \code{Gym} (gymnastics), \code{Netball}, \code{Rowing}, \code{Swim} (swimming), \code{T400m}, (track, further than 400m), \code{Tennis}, \code{TPSprnt} (track sprint events), \code{WPolo} (waterpolo)}
  \item{Ht}{The height, in cm}
  \item{Wt}{The weight of the athlete, in kg}
  \item{SSF}{The sum of skin folds}
  \item{PBF}{The percentage body fat}
}
}
\source{
OzDASL, available on-line at http://www.statsci.org/data/.
}
\usage{
data(AISsub)
}
\description{
Body measurements from athletes at the Australian Institute of Sport.
}
\references{
Telford, R. D. and Cunningham, R. B. (1991).
Sex, sport, and body-size dependency of hematology in highly trained athletes. 
\emph{Medicine and Science in Sports and Exercise}, \bold{23}(7):788–794.
}
\keyword{datasets}
