% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anorexia.R
\docType{data}
\name{Anorexia}
\alias{Anorexia}
\title{Weight loss after treatment for anorexia}
\format{
A data frame with 72 rows and 3 columns:
\describe{
  \item{Treatment}{The treatment type; one of \code{CB} (cognitive behavioural treatment), \code{Control} (the control group) or \code{FT} (family therapy)}
  \item{Before}{Weight (in kg) before the anorexia treatment}
  \item{After}{Weight (in kg) after the anorexia treatment}
}
}
\source{
Hand et al. \emph{A Handbook of Small Data Sets}, data set #285.
}
\usage{
data(Anorexia)
}
\description{
Weight changes in girls with anorexia: two treatments.
}
\keyword{datasets}
