% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B12Diet.R
\docType{data}
\name{B12Diet}
\alias{B12Diet}
\title{Vegetarianism and B12}
\format{
A data frame with 124 rows (one for each person) and 2 columns:
\describe{
  \item{B12}{B12 deficiency; one of \code{1} (B12 deficient) or \code{2} (Not B12 deficient)}
  \item{Diet}{The diet; one of \code{1} (Vegetarian) or `\code{2} (non-vegetarian)}
}
}
\source{
Gammon, Cheryl S., Pamela R. von Hurst, Joan Coad, Rozanne Kruger, and Welma Stonehouse. 2012.
Vegetarianism, Vitamin B12, and Insulin Resistance in a Group of Predominately
Overweight/Obese South Asian Women. \emph{Nutrition} \bold{28}: 20--24.
}
\usage{
data(B12Diet)
}
\description{
B12 deficiency in vegetarian and non-vegetarian women.
}
\keyword{datasets}
