% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BabyBoom.R
\docType{data}
\name{BabyBoom}
\alias{BabyBoom}
\title{Baby births in one day}
\format{
A data frame with 44 rows (one per birth) and 3 columns:
\describe{
  \item{Gender}{The gender of the child; one of \code{Female} or \code{Male}}
  \item{Weight}{The weight of the baby, in kg}
  \item{Mins.Since.Midnight}{the time of birth, in minutes since midnight}
}
}
\source{
Steele, S. 1997. Babies by the Dozen for Christmas: 24-Hour Baby Boom. \emph{The Sunday Mail}, 7.

Dunn, Peter K. 1999. A Simple Dataset for Demonstrating Common Distributions. 
\emph{Journal of Statistics Education}, \bold{7} (3).
}
\usage{
data(BabyBoom)
}
\description{
Details of the births on one day from a Brisbane hospital.
}
\keyword{datasets}
