% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Captopril.R
\docType{data}
\name{Captopril}
\alias{Captopril}
\title{Captopril}
\format{
A data frame with 30 rows (one per person) and 3 columns:
\describe{
  \item{Before}{The blood pressure before taking captopril, in mm Hg}
  \item{After}{The blood pressure after taking captopril, in mm Hg}
  \item{BP}{The type of blood pressure measured; \code{S} for systolic, and \code{D} for diastolic}
}
}
\source{
Hand et al. \emph{A Handbook of Small Data Sets}, data set #72.

Originally from: MacGregor, Graham A., N. D.Markandu, J. E. Roulston, and J. C. Jones. 1979. 
Essential Hypertension: Effect of an Oral Inhibitor of Angiotensin-Converting Enzyme.
\emph{British Medical Journal} \bold{2}: 1106--9.
}
\usage{
data(Captopril)
}
\description{
Blood pressure before and after treatment with Captopril.
}
\keyword{datasets}
