% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Corollas.R
\docType{data}
\name{Corollas}
\alias{Corollas}
\title{Price of second-hand Corollas}
\format{
A data frame with 45 rows (one per vehicle) and 3 columns:
\describe{
  \item{Year}{the year of manufacture of the vehicle}
  \item{Price}{the advertised price, in AUD}
  \item{Age}{the age of the vehicle, in years}
}
}
\source{
Collected by Peter K. Dunn, 2014, from www.gumtree.com.au

Peter K. Dunn (2023). \emph{Scientific Research and Methodology: An introduction to quantitative research in science and health}.
}
\usage{
data(Corollas)
}
\description{
The price of second-hand Corollas advertised on Gum Tree.
}
\keyword{datasets}
