% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deceleration.R
\docType{data}
\name{Deceleration}
\alias{Deceleration}
\title{Deceleration}
\format{
A data frame with 79 rows (one per person) and 2 columns:
\describe{
  \item{When}{When the deceleration is measured: \code{Before} or \code{After} signage added}
  \item{Deceleration}{The deceleratio}
}
}
\source{
Ma, Yongfeng, Wenbo Zhang, Xin Gu, and Jiguang Zhao. 2019.
Impacts of Experimental Advisory Exit Speed Sign on Traffic Speeds for Freeway Exit Ramp. 
\emph{PLoS One} \bold{14} (11): e0225203.
}
\usage{
data(Deceleration)
}
\description{
Deceleration data set
}
\keyword{datasets}
