% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDpatients.R
\docType{data}
\name{EDpatients}
\alias{EDpatients}
\title{EDpatients}
\format{
A data frame with 30 rows (one per person) and 2 columns:
\describe{
  \item{Days}{The number of days after welfare distribution}
  \item{ED}{The mean number of emergency department (ED) patients}
}
}
\source{
Data read from the scatterplot in 
Brunette, Douglas D., John Kominsky, and Ernest Ruiz. 1991. 
Correlation of Emergency Health Care Use, 911 Volume, and Jail Activity with Welfare Check Distribution.
\emph{Annals of Emergency Medicine} \bold{20} (7): 739--42.
}
\usage{
data(EDpatients)
}
\description{
Welfare distribution and emergency department patients.
}
\keyword{datasets}
