% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ferretin.R
\docType{data}
\name{Ferritin}
\alias{Ferritin}
\title{Ferritin}
\format{
A data frame with 20 rows (one per patient) and 3 columns:
\describe{
  \item{September}{The patients' ferritin content (in micrograms/L) in September}
  \item{March}{The patients' ferritin content (in micrograms/L) in March}
  \item{Reduction}{The reduction in the patients' ferritin content (in micrograms/L) between September and the following March, during which time they had treatment}
}
}
\source{
Cressie, N. A. C., L. J. Sheffield, and H. J.Whitford. 1984. 
Use of the One Sample $t$-Test in the Real World. 
\emph{Journal of Chronic Diseases} \bold{37} (2): 107--14.
}
\usage{
data(Ferritin)
}
\description{
Ferritin concentration changes.
}
\keyword{datasets}
