% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fluoro.R
\docType{data}
\name{Fluoro}
\alias{Fluoro}
\title{Fluoro}
\format{
A data frame with 19 rows and 2 columns:
\describe{
  \item{Time}{The total procedure time, in minutes}
  \item{Dose}{The total radiation dose received, in rads}
}
}
\source{
Kelly H. Zou, Kemal Tuncali, and Stuart G. Silverman (2003).
Correlation and simple linear regression. \emph{Radiology}, \bold{227}, 617–628.
}
\usage{
data(Fluoro)
}
\description{
The data give the total procedure time during ct fluoroscopic scanning, and the radiation dose received.
}
\references{
The data were originally used, but not given, in:
S. G. Silverman, K. Tuncali, D. F. Adams, R. D. Nawfel, K. H. Zou, and P. F. Judy (1999).
{CT} fluoroscopy-guided abdominal interventions: techniques, results, and radiation exposure. 
\emph{Radiology}, \bold{212}, 673–681.
}
\keyword{datasets}
