% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fruit.R
\docType{data}
\name{Fruit}
\alias{Fruit}
\title{Fruit}
\format{
A data frame with 37 rows (one per person) and 8 columns:
\describe{
  \item{Farm}{The farm identifier}
  \item{Flowers2014}{The number of flowers in 2014}
  \item{Flowers2015}{The number of flowers in 2015}
  \item{Fruit2014}{The total number of fruits formed in 2014}
  \item{Fruit2015}{The total number of fruits formed in 2015}
  \item{FLength2014}{The fruit length (in cm) in 2014}
  \item{FLength2015}{The fruit length (in cm) in 2015}
  \item{FBreadth2014}{The fruit breadth (in cm) in 2014}
  \item{FBreadth2015}{The fruit breadth (in cm) in 2015}
  \item{FWeight2014}{The fruit weight (in g) in 2014}
  \item{FWeight2015}{The fruit weight (in g) in 2015}
}
}
\source{
Mukherjee, Ronita; Deb, Rittik; Devy, Soubadra (2020). 
Data from: Diversity matters: effects of density compensation in pollination service during rainfall shift [Dataset]. 
Dryad. https://doi.org/10.5061/dryad.0n5v168
}
\usage{
data(Fruit)
}
\description{
Details of fruit from different farms.
}
\keyword{datasets}
