% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Insulation.R
\docType{data}
\name{Insulation}
\alias{Insulation}
\title{Insulation}
\format{
A data frame with 10 rows (each house) and 2 columns:
\describe{
  \item{Before}{Energy consumption before adding insulation, in MWh}
  \item{After}{Energy consumption after adding insulation, in MWh}
}
}
\source{
Hand et al. \emph{A Handbook of Small Data Sets}, data set #86.

Originally from: The Open University. 1983. \emph{MDST242 Statistics in Society, Unit A0: Introduction.} The Open University.
}
\usage{
data(Insulation)
}
\description{
Energy consumption before and after adding insulation.
}
\keyword{datasets}
