% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kerbside.R
\docType{data}
\name{Kerbside}
\alias{Kerbside}
\title{Kerbside}
\format{
A data frame with 6 rows (each combination) and 3 columns:
\describe{
  \item{Type}{Type of rubbish; one of \code{Reuseable} or \code{Non-reuseable}}
  \item{Opinion}{The person's opinion; one of \code{Acceptable}, \code{Not acceptable} or \code{Conditionally acceptable}}
  \item{Counts}{The number of people whose opinion aligns with the given combination}
}
}
\source{
Comerford, Emma, Jeanette Durante, Rod Goldsworthy, Vicki Hall, James Gooding, and Bernie Quinn. 2018.
Motivations for Kerbside Dumping: Evidence from Brisbane, Australia. 
\emph{Waste Management} \bold{78}: 490--96.
}
\usage{
data(Kerbside)
}
\description{
The Kerbside dataset
}
\keyword{datasets}
