% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lynx.R
\docType{data}
\name{Lynx}
\alias{Lynx}
\title{Lynx}
\format{
A data frame with 114 rows (each year) and 2 columns:
\describe{
  \item{Year}{The year}
  \item{Trappings}{The number of lynx trappings}
}
}
\source{
Elton, Charles, and Mary Nicholson. 1942. The Ten-Year Cycle in Numbers of the Lynx in Canada. 
\emph{The Journal of Animal Ecology}, 215--44.
Many have used and given this data.
}
\usage{
data(Lynx)
}
\description{
The Lynx dataset
}
\keyword{datasets}
