% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaryRiver.R
\docType{data}
\name{MaryRiver}
\alias{MaryRiver}
\title{Mary River stream flow data}
\format{
A data frame with 21,659 rows and 8 columns:
\describe{
  \item{Month}{The month (where \code{1} means January, etc.}
  \item{Year}{The year}
  \item{Mean}{The mean stream flow recording for given date, in ML}
}
}
\source{
The State of Queensland.
Originally sourced from <http://watermonitoring.dnrm.qld.gov.au/cgi/webhyd.pl?rsdf_org=138110A&cat=rs&lvl=1&0>,
but the actual website keeps changing...

Last time I checked it was <https://water-monitoring.information.qld.gov.au>;
then select "Streamflow data", "Mary Basin" and "Mary River at Bellbird Creek" (i.e., station 138110A).
}
\usage{
data(MaryRiver)
}
\description{
The mean daily stream flow from the Mary River.
}
\keyword{datasets}
