% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMiner.R
\docType{data}
\name{NMiner}
\alias{NMiner}
\title{NMiner}
\format{
A data frame with 31 rows (each foetus) and 8 columns:
\describe{
  \item{Eucs}{The number of eucalypt trees in the transect}
  \item{Minerab}{The number of noisy miners ('abundance') in three 20 min surveys in each transect}
}
}
\source{
Personal communication from Martine Maron.
}
\usage{
data(NMiner)
}
\description{
The number of noisy miners detected in various 2 hectare transects in buloke woodland patches within the Wimmera Plains of western Victoria, Australia
}
\references{
Martine Maron (2007) Threshold effect of eucalypt density on an aggressive avian competitor. 
\emph{Biological Conservation}, \bold{136}, 100–107.
}
\keyword{datasets}
