% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Punting.R
\docType{data}
\name{Punting}
\alias{Punting}
\title{Punting}
\format{
A data frame with 13 rows (each footballer) and 3 columns:
\describe{
  \item{Left}{The left-leg strength, in pounds}
  \item{Right}{The right-leg strength, in pounds}
  \item{Punt}{The punting distance, in feet}
}
}
\source{
Data in Exercise 2.11 of: 
Myers, Raymond H. 1990. \emph{Classical and Modern Regression with Applications}. Second edition. Duxbury.
}
\usage{
data(Punting)
}
\description{
The Punting dataset
}
\keyword{datasets}
