% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReactionTimePhoneInd.R
\docType{data}
\name{ReactionTimePhoneInd}
\alias{ReactionTimePhoneInd}
\title{ReactionTimePhoneInd}
\format{
A data frame with 64 rows (each student) and 2 columns:
\describe{
  \item{Reaction}{The reaction time, in milliseconds}
  \item{Group}{Which group the student was in; one of \code{Phone} or \code{Control}}
}
}
\source{
Reported by: 
Agresti, Alan, and Christine A. Franklin. 2007. 
\emph{Statistics: The Art and Science of Learning from Data}.

Agresti & Franklin claim the data comes from:
Strayer, David L., and William A. Johnston. 2001. 
Driven to Distraction: Dual-Task Studies of Simulated Driving and Conversing on a Cellular Telephone. 
\emph{Psychological Science} \bold{12} (6):462--66
}
\usage{
data(ReactionTimePhoneInd)
}
\description{
The ReactionTimePhoneInd dataset
}
\keyword{datasets}
