% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedDeer.R
\docType{data}
\name{RedDeer}
\alias{RedDeer}
\title{RedDeer}
\format{
A data frame with 78 rows (each deer) and 2 columns:
\describe{
  \item{Age}{The age of the deer, in years}
  \item{Weight}{The weight of the first molar tooth, in g}
}
}
\source{
Hand et al. \emph{A Handbook of Small Data Sets}, data set #170.

The data originally come from: 
Holgate, P. 1965. Fitting a Straight Line to Data from a Truncated Population.
\emph{Biometrics} \bold{21}(3): 715--20
}
\usage{
data(RedDeer)
}
\description{
The RedDeer dataset
}
\keyword{datasets}
