% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Removal.R
\docType{data}
\name{Removal}
\alias{Removal}
\title{Removal}
\format{
A data frame with 32 rows (each experiment) and 2 columns:
\describe{
  \item{Removal}{The removal efficieny, in percent}
  \item{Temp}{The inlet temperature, in degrees C}
}
}
\source{
Exercise 12.109 in 
Devore, Jay L., and Kenneth N. Berk. 2007. \emph{Modern Mathematical Statistics with Applications}.
Thomson Higher Education

The data originally come from: 
Chitwood, Derek E., and Joseph S. Devinny. 2001. 
Treatment of Mixed Hydrogen Sulfide and Organic Vapors in a Rock Medium Biofilter. 
\emph{Water Environment Research} \bold{73} (4): 426--35.
}
\usage{
data(Removal)
}
\description{
The removal efficiency in biofiltration.
}
\keyword{datasets}
