%% File Name: SRMData-package.Rd
%% File Version: 0.01

\name{SRMData-package}
\alias{SRMData-package}
\alias{SRMData}
\docType{package}

\title{
Datasets for the Social Relations Model (SRMData)
}

\description{
    Contains several original and resimulated datasets of 
    published articles for the social relations model.
}

%% \details{

%% The \pkg{LAM} package contains the following main functions: 

%% \itemize{   
%%
%% mlnormal
%% \item A general fitt
%%
%% amh, pmle
%%\item A general (but
%%
%% }

% }


\author{
Alexander Robitzsch, Oliver Luedtke
IPN - Leibniz Institute for Science and Mathematics Education
at Kiel University \cr \cr
Maintainer: Alexander Robitzsch <robitzsch@ipn.uni-kiel.de> \cr \cr
}


%% include references
%\references{
%Luedtke, O., Robitzsch, A., & Wagner, J. (xxxx). More stable estimation of the STARTS model:  
%}

%\seealso{
%XXXXXXX
%}

\keyword{package}
