% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stress.R
\docType{data}
\name{Stress}
\alias{Stress}
\title{Stress}
\format{
A data frame with 19 rows and 2 columns:
\describe{
  \item{BeforeHours}{beta-endorphin concentrations measured 12--14 hours before surgery, in fmol/ml}
  \item{BeforeMins}{beta-endorphin concentrations measured 10 minutes before surgery, in fmol/ml}
}
}
\source{
Hand et al. \emph{A Handbook of Small Data Sets}, data set #232.

The original source is given as Hoaglin, D. C., Mosteller, F. and Tukey. J. W. 1985.
\emph{Exploring data tables, trends and shapes}. New York: John Wiley & Sons.
}
\usage{
data(Stress)
}
\description{
Stress before surgery.
}
\keyword{datasets}
