% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudentsEat.R
\docType{data}
\name{StudentsEat}
\alias{StudentsEat}
\title{StudentsEat}
\format{
A data frame with 183 rows (each student) and 2 columns:
\describe{
  \item{Meals}{Where the student eats most of their meals; one of \code{Most off-campus} or \code{Most on-campus}}
  \item{Live}{Where the student lives; one of \code{Living with parents} or \code{Not living with parents}}
}
}
\source{
Mann, Linda, and Karen Blotnicky. 2017.
Influences of Physical Environments on University Student Eating Behaviors.
\emph{International Journal of Health Sciences} \bold{5} (2): 42–52
}
\usage{
data(StudentsEat)
}
\description{
Where students live and where they eat most of their meals.
}
\keyword{datasets}
