% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudentWt.R
\docType{data}
\name{StudentWt}
\alias{StudentWt}
\title{StudentWt}
\format{
A data frame with 68 rows (each student) and 4 columns:
\describe{
  \item{Student}{An identifier}
  \item{Week1}{The student's weight in Week 1, in kg}
  \item{Week12}{The student's weight in Week 12, in kg}
  \item{GainWt}{The student's weight gain, in kg}
}
}
\source{
Levitsky, D. A., Halbmaier, C. A., & Mrdjenovic, G. (2004).
The freshman weight gain: a model for the study of the epidemic of obesity. 
\emph{International Journal of Obesity}, \bold{28}(11), 1435--1442.

The data are given at:  David. n.d. DASL: Data and Story Library. <https://dasl.datadescription.com/datafile/freshman-15/>
}
\usage{
data(StudentWt)
}
\description{
Weights of students from Week 1 to Week 12 of semester.
}
\keyword{datasets}
