% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YieldDen.R
\docType{data}
\name{YieldDen}
\alias{YieldDen}
\title{YieldDen}
\format{
A data frame with 30 rows (each plants) and 3 columns:
\describe{
  \item{Yield}{The yield per plant, in grams}
  \item{Dens}{The planting density, in plants per square foot}
  \item{Var}{The variety; one of \code{1}. \code{2} or \code{3}}
}
}
\source{
R. Mead (1970). Plant density and crop yield. 
\emph{Applied Statistics}, \bold{19}(1), 64–81.
}
\usage{
data(YieldDen)
}
\description{
The mean yields per plant for three onion varieties
}
\keyword{datasets}
