%% File Name: SRMData-package.Rd
%% File Version: 0.03

\name{SRMData-package}
\alias{SRMData-package}
\alias{SRMData}
\docType{package}

\title{
\packageTitle{SRMData}
}

\description{
\packageDescription{SRMData}
}

%% \details{

%% The \pkg{LAM} package contains the following main functions: 

%% \itemize{   
%%
%% mlnormal
%% \item A general fitt
%%
%% amh, pmle
%%\item A general (but
%%
%% }

% }


\author{
\packageAuthor{SRMData}

Maintainer: \packageMaintainer{SRMData}
}


%% include references
%\references{
%Luedtke, O., Robitzsch, A., & Wagner, J. (xxxx). More stable estimation of the STARTS model:  
%}

%\seealso{
%XXXXXXX
%}

\keyword{package}
